backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
rank-bm25==0.2.2
groq==0.15.0
numpy==1.26.4
